+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * **BREAKING** Removed the `ResponseQueue` in favor of `:telemetry` execution
 * Notably adjust `duration` as determined in the `Honeylixir.Response` object to be a floating point number of milliseconds, such that something taking 100 microseconds is represented roughly as 0.1.
 * Add support for Global field configuration
+* No longer drop events when `max_children` is reached and no more `TransmissionSender`s can be started
 
 # 0.3.0